pyproject.toml
[tool.poetry]
name = "prophecies"
version = "0.4.0-alpha0"
authors = ['ICIJ <engineering@icij.org>']
description = "An ICIJ app to conduct data validation and cleaning"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "==3.2.16"
djangorestframework = "3.13.1"
django-cors-headers = "*"
django-debug-toolbar = "*"
django-environ = "*"
social-auth-app-django = "*"
django-webpack-loader = "*"
psycopg2-binary = "*"
gunicorn = "*"
pillow = "*"
django-colorfield = "*"
django-constance = {extras = [ "database",], version = "2.9.1"}
django-filter = "*"
django-crispy-forms = "*"
djangorestframework-jsonapi = {extras = ["django-polymorphic"], version = "*"}
django-activity-stream = "1.4.1"
django-jsonfield = "*"
markdown = "*"
django-admin-autocomplete-filter = "*"
django-import-export = "*"

[tool.poetry.group.dev.dependencies]
time-machine = "*"
pytest-django = "*"
pylint = "^2.17.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
